Read an input text from the console and print its characters in reversed order.
Use Reader for collecting characters.

In this task, the input is limited by 50 characters. However, you are welcome to find a solution that does not depend on the input size, which may require some extra knowledge.

It is important to understand the difference between the input stream closing and pressing Enter in case of testing in IDEA. Hyperskill platform produces test input and sends end-of-file event right away. It does not append additional end-of-line symbols.
Hint

Sample Input:
abc

Sample Output:
cba